requirements.txt
fredapi==0.5.2
python-dotenv==1.1.1
pandas
